feat: add CustomerDeliveryDetailMovementCard component
